Improved line editing with lined.
Added support for escape (meta) sequences. This involved
changing the ose_termRead function to take a buffer and max
number of chars to read. Ultimately, the reading is happening
in the wrong place--it should be done by the host, and
ose_lined_read should read things from the stack, not a file
descriptor. To be refactored in a future commit.

Added M^b (jump back), M^f (jump forward), and M^DEL (delete word).
In this commit, the only word-break character is '/', and it's
hardcoded. In a future commit, we'll use a list of chars that's
setable.

Also added C^d for deleting the character under the cursor.